Achieve ansible provision in EC2 AWS
This playbook is able to launch and provision one or several ec2 machines, also, assign them private ip's.

As the aws_ec2.yml file shows, the register enable to save the result of the ec2 task, however, is it necessary to know how to iterate over it, as the result value could be a complex format.

The format of this ec2 variable contains a map with a 'result' field, the type of this latest variable is a list, where each element could contains another map, where finally there are instances of the ec2 machines data