Change hostname of each host for operability and reliability
In one run using a non-default security group in aws, an exception was launched due the fact that the hostname did not match any alias in /etc/hosts file. So, in order to avoid this kind of issue, each machine (master or slave) has his own hostname matching an alias in /etc/hosts file.

inventory_hostname variable contains a default and generated identifier of the current machine, and play_hosts.index(inventory_hostname) obtains the index of that machine in the group.